Allow choosing branch name when open in browser
This way, I can have a keymap to open link using `master` branch, and
another keymap to open using the current branch. Most of the time what I
want is using `master`.

Line numbers when changing branch this way might be incorrect, but it's
not a big problem (for now).